Modified nearly all the places where the "raw" Address() constructor was being used to use one of the following ctors: * Address.Ptr16 * Address.Ptr32 * Address.Ptr64 * Address.SegPtr
A pleasant side effect is 64- and 16-bit addresses appearing correctly in LowLevelViewer.

+++ src/Arch/Z80/Z80Disassembler.cs
@@ -118,7 +118,7 @@ namespace Decompiler.Arch.Z80
                 case 'J':       // Relative jump
                     var width = OperandSize(fmt[i++]);
                     int ipOffset = rdr.ReadLeSigned(width);
-                    ops[iOp++] = AddressOperand.Ptr16((uint)(rdr.Address.Offset + ipOffset));
+                    ops[iOp++] = AddressOperand.Ptr16((ushort)(rdr.Address.Offset + ipOffset));
                     break;
                 case 'x':       // 2-digit Inline hexadecimal byte
                     int val = (Hex(fmt[i++]) << 4);
